feat. Create function
First, the ciphertext function takes a plaintext p and a number of runs k as parameters.

Then define the ciphertext to be solved in C.

It then returns a loop of the length of the plaintext.

The iterator uses the ord function in the plaintext p argument to get the ASCII value.

It then adds +K to the ASCII value and adds it to the ciphertext string C.

Here, if the plaintext contains a space (ASCII value 32), we replace it with 64.

This is because we don't use ASCII values 33 through 64, and we move on to the A value (65) after the jump (32).

We also go backwards again after the Z value (90), so if the ASCII value is greater than 90, we give it a -27 so that we can go back to the previous character.

to get us back to the previous character.